refactor: Move EdgeData to types
