+++ packages/mermaid/src/diagrams/class/classTypes.ts
@@ -137,24 +137,6 @@ export interface ClassNote {
   text: string;
 }
 
-export interface EdgeData {
-  arrowheadStyle?: string;
-  labelpos?: string;
-  labelType?: string;
-  label?: string;
-  classes: string;
-  pattern: string;
-  id: string;
-  arrowhead: string;
-  startLabelRight: string;
-  endLabelLeft: string;
-  arrowTypeStart: string;
-  arrowTypeEnd: string;
-  style: string;
-  labelStyle: string;
-  curve: any;
-}
-
 export type ClassRelation = {
   id1: string;
   id2: string;
